test: move lua option/variable tests to a separate file
